requirements.txt
future==0.18.2
pip~=22.3.1
wheel~=0.37.1
setuptools~=65.4.1
python-dateutil~=2.8.2